+++ src/main/java/com/nemonotfound/nemos/creatures/NemosCreatures.java
@@ -6,6 +6,10 @@ import com.nemonotfound.nemos.creatures.entity.CreaturesEntityTypes;
 import com.nemonotfound.nemos.creatures.item.CreaturesItemGroups;
 import com.nemonotfound.nemos.creatures.item.CreaturesItems;
 import net.fabricmc.api.ModInitializer;
+import net.fabricmc.fabric.api.loot.v3.LootTableEvents;
+import net.minecraft.entity.EntityType;
+import net.minecraft.loot.LootPool;
+import net.minecraft.loot.entry.ItemEntry;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -26,6 +30,15 @@ public class NemosCreatures implements ModInitializer {
 		CreaturesItems.bootstrap();
 		CreaturesItemGroups.bootstrap();
 		ModDispenserBehavior.registerDefaults();
+
+		LootTableEvents.MODIFY.register((key, tableBuilder, source, wrapperLookup) -> {
+			if (source.isBuiltin() && EntityType.PARCHED.getLootTableKey().isPresent() && EntityType.PARCHED.getLootTableKey().get() == key) {
+				var pool = LootPool.builder()
+						.with(ItemEntry.builder(CreaturesItems.PARCHED_BONE));
+
+				tableBuilder.pool(pool);
+			}
+		});
 	}