+++ src/main/java/com/nemonotfound/nemos/creatures/datagen/EntityTypeTagsProvider.java
@@ -1,12 +1,14 @@
 package com.nemonotfound.nemos.creatures.datagen;
 
-import com.nemonotfound.nemos.creatures.entity.CreaturesEntityTypes;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricTagProvider;
 import net.minecraft.registry.RegistryWrapper;
 import net.minecraft.registry.tag.EntityTypeTags;
+import org.jspecify.annotations.NonNull;
 
 import java.util.concurrent.CompletableFuture;
+
+import static com.nemonotfound.nemos.creatures.entity.CreaturesEntityTypes.*;
 
 public class EntityTypeTagsProvider extends FabricTagProvider.EntityTypeTagProvider {
 
@@ -15,13 +17,27 @@ public class EntityTypeTagsProvider extends FabricTagProvider.EntityTypeTagProvi
     }
 
     @Override
-    protected void configure(RegistryWrapper.WrapperLookup wrapperLookup) {
+    protected void configure(RegistryWrapper.@NonNull WrapperLookup wrapperLookup) {
         valueLookupBuilder(EntityTypeTags.SKELETONS)
-                .add(CreaturesEntityTypes.CRIMSON_SKELETON)
-                .add(CreaturesEntityTypes.SCORCHED_SKELETON)
-                .add(CreaturesEntityTypes.SNOWY_SKELETON)
-                .add(CreaturesEntityTypes.FROZEN_SKELETON)
-                .add(CreaturesEntityTypes.VENOMOUS_SKELETON)
-                .add(CreaturesEntityTypes.WARPED_SKELETON);
+                .add(CRIMSON_SKELETON)
+                .add(SNOWY_SKELETON)
+                .add(FROZEN_SKELETON)
+                .add(VENOMOUS_SKELETON)
+                .add(WARPED_SKELETON);
+
+        valueLookupBuilder(EntityTypeTags.ZOMBIES)
+                .add(SNOWY_ZOMBIE)
+                .add(FROZEN_ZOMBIE)
+                .add(VENOMOUS_ZOMBIE);
+
+        valueLookupBuilder(EntityTypeTags.BURN_IN_DAYLIGHT)
+                .add(CRIMSON_SKELETON)
+                .add(SNOWY_SKELETON)
+                .add(FROZEN_SKELETON)
+                .add(VENOMOUS_SKELETON)
+                .add(WARPED_SKELETON)
+                .add(SNOWY_ZOMBIE)
+                .add(FROZEN_ZOMBIE)
+                .add(VENOMOUS_ZOMBIE);
     }
 }