+++ src/main/java/com/nemonotfound/nemos/creatures/entity/CreaturesEntityTypes.java
@@ -23,88 +23,141 @@ import static com.nemonotfound.nemos.creatures.NemosCreatures.MOD_ID;
 
 public class CreaturesEntityTypes {
 
-    public static final EntityType<VenomousSkeletonEntity> VENOMOUS_SKELETON = registerEntityType("venomous_skeleton",
+    public static final EntityType<VenomousSkeletonEntity> VENOMOUS_SKELETON = register(
+            "venomous_skeleton",
             EntityType.Builder.create(VenomousSkeletonEntity::new, SpawnGroup.MONSTER)
-                    .dimensions(0.6f, 1.99f));
-    public static final EntityType<VenomousSpiderEntity> VENOMOUS_SPIDER = registerEntityType("venomous_spider",
+                    .dimensions(0.6f, 1.99f)
+                    .eyeHeight(1.74F)
+                    .vehicleAttachment(-0.7F)
+                    .maxTrackingRange(8)
+                    .notAllowedInPeaceful()
+    );
+    public static final EntityType<VenomousSpiderEntity> VENOMOUS_SPIDER = register(
+            "venomous_spider",
             EntityType.Builder.create(VenomousSpiderEntity::new, SpawnGroup.MONSTER)
-                    .dimensions(0.8f, 0.6f));
-    public static final EntityType<VenomousZombieEntity> VENOMOUS_ZOMBIE = registerEntityType("venomous_zombie",
+                    .dimensions(0.8f, 0.6f)
+                    .notAllowedInPeaceful());
+    public static final EntityType<VenomousZombieEntity> VENOMOUS_ZOMBIE = register(
+            "venomous_zombie",
             EntityType.Builder.create(VenomousZombieEntity::new, SpawnGroup.MONSTER)
-                    .dimensions(0.6f, 1.95f));
-    public static final EntityType<VenomousCreeperEntity> VENOMOUS_CREEPER = registerEntityType("venomous_creeper",
+                    .dimensions(0.6f, 1.95f)
+                    .notAllowedInPeaceful()
+    );
+    public static final EntityType<VenomousCreeperEntity> VENOMOUS_CREEPER = register(
+            "venomous_creeper",
             EntityType.Builder.create(VenomousCreeperEntity::new, SpawnGroup.MONSTER)
-                    .dimensions(0.6f, 1.7f));
-    public static final EntityType<ScorchedSkeletonEntity> SCORCHED_SKELETON = registerEntityType("scorched_skeleton",
-            EntityType.Builder.create(ScorchedSkeletonEntity::new, SpawnGroup.MONSTER)
-                    .dimensions(0.6f, 1.99f));
-    public static final EntityType<SandSpiderEntity> SAND_SPIDER = registerEntityType("sand_spider",
+                    .dimensions(0.6f, 1.7f)
+                    .notAllowedInPeaceful()
+    );
+    public static final EntityType<SandSpiderEntity> SAND_SPIDER = register(
+            "sand_spider",
             EntityType.Builder.create(SandSpiderEntity::new, SpawnGroup.MONSTER)
-                    .dimensions(0.4f, 0.2f));
-    public static final EntityType<MummyEntity> MUMMY = registerEntityType("mummy",
-            EntityType.Builder.create(MummyEntity::new, SpawnGroup.MONSTER)
-                    .dimensions(0.6f, 1.95f));
-    public static final EntityType<ScorchedCreeperEntity> SCORCHED_CREEPER = registerEntityType("scorched_creeper",
+                    .dimensions(0.4f, 0.2f)
+                    .notAllowedInPeaceful()
+    );
+    public static final EntityType<ScorchedCreeperEntity> SCORCHED_CREEPER = register(
+            "scorched_creeper",
             EntityType.Builder.create(ScorchedCreeperEntity::new, SpawnGroup.MONSTER)
-                    .dimensions(0.6f, 1.7f));
-    public static final EntityType<CrimsonSkeletonEntity> CRIMSON_SKELETON = registerEntityType("crimson_skeleton",
+                    .dimensions(0.6f, 1.7f)
+                    .notAllowedInPeaceful()
+    );
+    public static final EntityType<CrimsonSkeletonEntity> CRIMSON_SKELETON = register(
+            "crimson_skeleton",
             EntityType.Builder.create(CrimsonSkeletonEntity::new, SpawnGroup.MONSTER)
-                    .dimensions(0.6f, 1.99f));
-    public static final EntityType<WarpedSkeletonEntity> WARPED_SKELETON = registerEntityType("warped_skeleton",
+                    .dimensions(0.6f, 1.99f)
+                    .notAllowedInPeaceful()
+    );
+    public static final EntityType<WarpedSkeletonEntity> WARPED_SKELETON = register(
+            "warped_skeleton",
             EntityType.Builder.create(WarpedSkeletonEntity::new, SpawnGroup.MONSTER)
-                    .dimensions(0.6f, 1.99f));
-    public static final EntityType<WildBoarEntity> WILD_BOAR = registerEntityType("wild_boar",
+                    .dimensions(0.6f, 1.99f)
+                    .notAllowedInPeaceful()
+    );
+    public static final EntityType<WildBoarEntity> WILD_BOAR = register("wild_boar",
             EntityType.Builder.create(WildBoarEntity::new, SpawnGroup.CREATURE)
-                    .dimensions(0.9f, 0.9f));
-    public static final EntityType<SnowySkeletonEntity> SNOWY_SKELETON = registerEntityType("snowy_skeleton",
+                    .dimensions(0.9f, 0.9f)
+    );
+    public static final EntityType<SnowySkeletonEntity> SNOWY_SKELETON = register(
+            "snowy_skeleton",
             EntityType.Builder.create(SnowySkeletonEntity::new, SpawnGroup.MONSTER)
-                    .dimensions(0.6f, 1.99f));
-    public static final EntityType<SnowySpiderEntity> SNOWY_SPIDER = registerEntityType("snowy_spider",
+                    .dimensions(0.6f, 1.99f)
+                    .notAllowedInPeaceful()
+    );
+    public static final EntityType<SnowySpiderEntity> SNOWY_SPIDER = register(
+            "snowy_spider",
             EntityType.Builder.create(SnowySpiderEntity::new, SpawnGroup.MONSTER)
                     .dimensions(1.4F, 0.9F)
                     .eyeHeight(0.65F)
                     .passengerAttachments(0.765F)
-                    .maxTrackingRange(8));
-    public static final EntityType<SnowSpiderEntity> SNOW_SPIDER = registerEntityType("snow_spider",
+                    .maxTrackingRange(8)
+                    .notAllowedInPeaceful()
+    );
+    public static final EntityType<SnowSpiderEntity> SNOW_SPIDER = register(
+            "snow_spider",
             EntityType.Builder.create(SnowSpiderEntity::new, SpawnGroup.MONSTER)
-                    .dimensions(0.4f, 0.2f));
-    public static final EntityType<SnowyCreeperEntity> SNOWY_CREEPER = registerEntityType("snowy_creeper",
+                    .dimensions(0.4f, 0.2f)
+                    .notAllowedInPeaceful()
+    );
+    public static final EntityType<SnowyCreeperEntity> SNOWY_CREEPER = register(
+            "snowy_creeper",
             EntityType.Builder.create(SnowyCreeperEntity::new, SpawnGroup.MONSTER)
                     .dimensions(0.6f, 1.7f));
-    public static final EntityType<SnowyZombieEntity> SNOWY_ZOMBIE = registerEntityType("snowy_zombie",
+    public static final EntityType<SnowyZombieEntity> SNOWY_ZOMBIE = register(
+            "snowy_zombie",
             EntityType.Builder.create(SnowyZombieEntity::new, SpawnGroup.MONSTER)
-                    .dimensions(0.6f, 1.95f));
-    public static final EntityType<SnowyPigEntity> SNOWY_PIG = registerEntityType("snowy_pig",
+                    .dimensions(0.6f, 1.95f)
+                    .notAllowedInPeaceful()
+    );
+    public static final EntityType<SnowyPigEntity> SNOWY_PIG = register(
+            "snowy_pig",
             EntityType.Builder.create(SnowyPigEntity::new, SpawnGroup.CREATURE)
                     .dimensions(0.9f, 0.9f)
                     .passengerAttachments(0.86875f)
-                    .maxTrackingRange(10));
-    public static final EntityType<SnowyCowEntity> SNOWY_COW =  registerEntityType("snowy_cow",
+                    .maxTrackingRange(10)
+    );
+    public static final EntityType<SnowyCowEntity> SNOWY_COW =  register(
+            "snowy_cow",
             EntityType.Builder.create(SnowyCowEntity::new, SpawnGroup.CREATURE)
                     .dimensions(0.9f, 1.4f)
                     .eyeHeight(1.3f)
                     .passengerAttachments(1.36875f)
-                    .maxTrackingRange(10));
-    public static final EntityType<FrozenSkeletonEntity> FROZEN_SKELETON = registerEntityType("frozen_skeleton",
+                    .maxTrackingRange(10)
+    );
+    public static final EntityType<FrozenSkeletonEntity> FROZEN_SKELETON = register(
+            "frozen_skeleton",
             EntityType.Builder.create(FrozenSkeletonEntity::new, SpawnGroup.MONSTER)
-                    .dimensions(0.6f, 1.99f));
-    public static final EntityType<FrozenCreeperEntity> FROZEN_CREEPER = registerEntityType("frozen_creeper",
+                    .dimensions(0.6f, 1.99f)
+                    .notAllowedInPeaceful()
+    );
+    public static final EntityType<FrozenCreeperEntity> FROZEN_CREEPER = register(
+            "frozen_creeper",
             EntityType.Builder.create(FrozenCreeperEntity::new, SpawnGroup.MONSTER)
-                    .dimensions(0.6f, 1.7f));
-    public static final EntityType<FrozenSpiderEntity> FROZEN_SPIDER = registerEntityType("frozen_spider",
+                    .dimensions(0.6f, 1.7f)
+                    .notAllowedInPeaceful()
+    );
+    public static final EntityType<FrozenSpiderEntity> FROZEN_SPIDER = register(
+            "frozen_spider",
             EntityType.Builder.create(FrozenSpiderEntity::new, SpawnGroup.MONSTER)
                     .dimensions(1.4F, 0.9F)
                     .eyeHeight(0.65F)
                     .passengerAttachments(0.765F)
-                    .maxTrackingRange(8));
-    public static final EntityType<IceSpiderEntity> ICE_SPIDER = registerEntityType("ice_spider",
+                    .maxTrackingRange(8)
+                    .notAllowedInPeaceful()
+    );
+    public static final EntityType<IceSpiderEntity> ICE_SPIDER = register(
+            "ice_spider",
             EntityType.Builder.create(IceSpiderEntity::new, SpawnGroup.MONSTER)
-                    .dimensions(0.4f, 0.2f));
-    public static final EntityType<FrozenZombieEntity> FROZEN_ZOMBIE = registerEntityType("frozen_zombie",
+                    .dimensions(0.4f, 0.2f)
+                    .notAllowedInPeaceful()
+    );
+    public static final EntityType<FrozenZombieEntity> FROZEN_ZOMBIE = register(
+            "frozen_zombie",
             EntityType.Builder.create(FrozenZombieEntity::new, SpawnGroup.MONSTER)
-                    .dimensions(0.6f, 1.95f));
+                    .dimensions(0.6f, 1.95f)
+                    .notAllowedInPeaceful()
+    );
 
-    public static <T extends Entity> EntityType<T> registerEntityType(String path, EntityType.Builder<T> entityTypeBuilder) {
+    public static <T extends Entity> EntityType<T> register(String path, EntityType.Builder<T> entityTypeBuilder) {
         Identifier id = Identifier.of(MOD_ID, path);
         RegistryKey<EntityType<?>> key = RegistryKey.of(RegistryKeys.ENTITY_TYPE, id);
 
@@ -116,9 +169,7 @@ public class CreaturesEntityTypes {
         SpawnRestriction.register(VENOMOUS_SPIDER, SpawnLocationTypes.ON_GROUND, Heightmap.Type.MOTION_BLOCKING, HostileEntity::canSpawnInDark);
         SpawnRestriction.register(VENOMOUS_ZOMBIE, SpawnLocationTypes.ON_GROUND, Heightmap.Type.MOTION_BLOCKING_NO_LEAVES, HostileEntity::canSpawnInDark);
         SpawnRestriction.register(VENOMOUS_CREEPER, SpawnLocationTypes.ON_GROUND, Heightmap.Type.MOTION_BLOCKING_NO_LEAVES, HostileEntity::canSpawnInDark);
-        SpawnRestriction.register(SCORCHED_SKELETON, SpawnLocationTypes.ON_GROUND, Heightmap.Type.MOTION_BLOCKING_NO_LEAVES, HostileEntity::canSpawnInDark);
         SpawnRestriction.register(SAND_SPIDER, SpawnLocationTypes.ON_GROUND, Heightmap.Type.MOTION_BLOCKING, HostileEntity::canSpawnInDark);
-        SpawnRestriction.register(MUMMY, SpawnLocationTypes.ON_GROUND, Heightmap.Type.MOTION_BLOCKING_NO_LEAVES, HostileEntity::canSpawnInDark);
         SpawnRestriction.register(SCORCHED_CREEPER, SpawnLocationTypes.ON_GROUND, Heightmap.Type.MOTION_BLOCKING_NO_LEAVES, HostileEntity::canSpawnInDark);
         SpawnRestriction.register(CRIMSON_SKELETON, SpawnLocationTypes.ON_GROUND, Heightmap.Type.MOTION_BLOCKING_NO_LEAVES, HostileEntity::canSpawnInDark);
         SpawnRestriction.register(WARPED_SKELETON, SpawnLocationTypes.ON_GROUND, Heightmap.Type.MOTION_BLOCKING_NO_LEAVES, HostileEntity::canSpawnInDark);
@@ -142,9 +193,7 @@ public class CreaturesEntityTypes {
         FabricDefaultAttributeRegistry.register(VENOMOUS_SPIDER, VenomousSpiderEntity.createJungleSpiderAttributes());
         FabricDefaultAttributeRegistry.register(VENOMOUS_ZOMBIE, VenomousZombieEntity.createZombieAttributes());
         FabricDefaultAttributeRegistry.register(VENOMOUS_CREEPER, VenomousCreeperEntity.createCreeperAttributes());
-        FabricDefaultAttributeRegistry.register(SCORCHED_SKELETON, ScorchedSkeletonEntity.createAbstractSkeletonAttributes());
         FabricDefaultAttributeRegistry.register(SAND_SPIDER, SandSpiderEntity.createSandSpiderAttributes());
-        FabricDefaultAttributeRegistry.register(MUMMY, MummyEntity.createMummyAttributes());
         FabricDefaultAttributeRegistry.register(SCORCHED_CREEPER, ScorchedCreeperEntity.createCreeperAttributes());
         FabricDefaultAttributeRegistry.register(CRIMSON_SKELETON, CrimsonSkeletonEntity.createAbstractSkeletonAttributes());
         FabricDefaultAttributeRegistry.register(WARPED_SKELETON, WarpedSkeletonEntity.createAbstractSkeletonAttributes());
@@ -177,9 +226,7 @@ public class CreaturesEntityTypes {
         BiomeModifications.addSpawn(jungle, SpawnGroup.MONSTER, VENOMOUS_SPIDER, 100, 4, 4);
         BiomeModifications.addSpawn(jungle, SpawnGroup.MONSTER, VENOMOUS_ZOMBIE, 95, 4, 4);
         BiomeModifications.addSpawn(jungle, SpawnGroup.MONSTER, VENOMOUS_CREEPER, 100, 4, 4);
-        BiomeModifications.addSpawn(desert, SpawnGroup.MONSTER, SCORCHED_SKELETON, 100, 4, 4);
         BiomeModifications.addSpawn(desert, SpawnGroup.MONSTER, SAND_SPIDER, 100, 4, 4);
-        BiomeModifications.addSpawn(desert, SpawnGroup.MONSTER, MUMMY, 100, 4, 4);
         BiomeModifications.addSpawn(desert, SpawnGroup.MONSTER, SCORCHED_CREEPER, 100, 4, 4);
         BiomeModifications.addSpawn(crimsonForest, SpawnGroup.MONSTER, CRIMSON_SKELETON, 40, 1, 1);
         BiomeModifications.addSpawn(warpedForest, SpawnGroup.MONSTER, WARPED_SKELETON, 60, 1, 1);