+++ changelog.md
@@ -1,4 +1,11 @@
-# Changelog v1.10.2
+# Changelog v2.0
+
+## Breaking Changes
+- Removed mobs since they're too similar to the parched
+  - Removed Mummy
+  - Removed Scorched Skeleton
+- Renamed scorched bone, bone meal and bone block to parched bone, bone meal and bone block
 
 ## Changes
-- Warped Skeleton shoots invisible arrows now
+- Parched now drop Parched Bone
+- Crimson, Warped and Frozen Skeleton now drop regular bones as well